Refactored quickslots into UiWindowController
