--- Emerald/Assets/Scripts/UiControllers/QuickSlotController.cs
@@ -1,59 +1,0 @@
-﻿using Emerald.UiControllers;
-using UnityEngine;
-using UnityEngine.InputSystem;
-
-public class QuickSlotController : MonoBehaviour {
-    private InputController.QuickSlotsActions quickSlotsActions;
-    private IReplaceMe[] quickSlotsEquipped;
-
-    private void Start() {
-        quickSlotsEquipped = new IReplaceMe[24];
-        quickSlotsActions = new InputController().QuickSlots;
-        quickSlotsActions.QuickSlot_F1.performed += callBack => StartQuickSlotAction((int)QuickSlot.F1);
-        quickSlotsActions.QuickSlot_F2.performed += callBack => StartQuickSlotAction((int)QuickSlot.F2);
-        quickSlotsActions.QuickSlot_F3.performed += callBack => StartQuickSlotAction((int)QuickSlot.F3);
-        quickSlotsActions.QuickSlot_F4.performed += callBack => StartQuickSlotAction((int)QuickSlot.F4);
-        quickSlotsActions.QuickSlot_F5.performed += callBack => StartQuickSlotAction((int)QuickSlot.F5);
-        quickSlotsActions.QuickSlot_F6.performed += callBack => StartQuickSlotAction((int)QuickSlot.F6);
-        quickSlotsActions.QuickSlot_S.performed += callBack => StartQuickSlotAction((int)QuickSlot.S);
-        quickSlotsActions.QuickSlot_D.performed += callBack => StartQuickSlotAction((int)QuickSlot.D);
-        quickSlotsActions.QuickSlot_Q.performed += callBack => StartQuickSlotAction((int)QuickSlot.Q);
-        quickSlotsActions.QuickSlot_Y.performed += callBack => StartQuickSlotAction((int)QuickSlot.Y);
-        quickSlotsActions.QuickSlot_B.performed += callBack => StartQuickSlotAction((int)QuickSlot.B);
-        quickSlotsActions.QuickSlot_H.performed += callBack => StartQuickSlotAction((int)QuickSlot.H);
-        quickSlotsActions.QuickSlot_1.performed += callBack => StartQuickSlotAction((int)QuickSlot.ONE);
-        quickSlotsActions.QuickSlot_2.performed += callBack => StartQuickSlotAction((int)QuickSlot.TWO);
-        quickSlotsActions.QuickSlot_3.performed += callBack => StartQuickSlotAction((int)QuickSlot.THREE);
-        quickSlotsActions.QuickSlot_4.performed += callBack => StartQuickSlotAction((int)QuickSlot.FOUR);
-        quickSlotsActions.QuickSlot_5.performed += callBack => StartQuickSlotAction((int)QuickSlot.FIVE);
-        quickSlotsActions.QuickSlot_6.performed += callBack => StartQuickSlotAction((int)QuickSlot.SIX);
-        quickSlotsActions.QuickSlot_7.performed += callBack => StartQuickSlotAction((int)QuickSlot.SEVEN);
-        quickSlotsActions.QuickSlot_8.performed += callBack => StartQuickSlotAction((int)QuickSlot.EIGHT);
-        quickSlotsActions.QuickSlot_9.performed += callBack => StartQuickSlotAction((int)QuickSlot.NINE);
-        quickSlotsActions.QuickSlot_0.performed += callBack => StartQuickSlotAction((int)QuickSlot.TEN);
-        quickSlotsActions.QuickSlot_minus.performed += callBack => StartQuickSlotAction((int)QuickSlot.MINUS);
-        quickSlotsActions.QuickSlot_equals.performed += callBack => StartQuickSlotAction((int)QuickSlot.EQUALS);
-    }
-
-    private void StartQuickSlotAction(int position) {
-        if (quickSlotsEquipped[position] == null) return;
-        quickSlotsEquipped[position].DoAction();
-    }
-
-    private void SetQuickSlot(QuickSlot position, IReplaceMe newItem) {
-        quickSlotsEquipped[(int) position] = newItem;
-    }
-
-    private void RemoveFromQuickSlot(QuickSlot position) {
-        quickSlotsEquipped[(int)position] = null;
-    }
-
-    public InputActionMap GetQuickSlotActions() => quickSlotsActions.Get();
-    public void EnableControls() => quickSlotsActions.Enable();
-    public void DisableControls() => quickSlotsActions.Disable();
-    public bool IsEnabled => quickSlotsActions.enabled;
-}
-
-internal interface IReplaceMe {
-    void DoAction();
-}
