--- Emerald/Assets/Scripts/UiControllers/QuickSlotController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 556976799773b4eb3a8068eb50cd58b2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
